frontend/capture_demo/requirements.txt
google-genai>=0.5.0 